requirements.txt
gspread
oauth2client
yagmail
python-dotenv
